+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/Library/MotionCaptureLoader/edit/bvh_remapangles.c
@@ -6,7 +6,7 @@
 
 /*
 //THIS IS NOT USED ANYWHERE
-double bvh_constrainAngleCentered180(double angle)
+float bvh_constrainAngleCentered180(float angle)
 {
    angle = fmod(angle,360.0);
    if (angle<0.0)
